Fast Quantile Regression

Fast Quantile Regression

fit_fqr(
  X,
  y,
  tau,
  se = T,
  init_beta = stats::rnorm(ncol(X)),
  smoothing_window = .Machine$double.eps,
  maxiter = 10000,
  beta_tol = 1e-16,
  check_tol = 1e-15 * nrow(X),
  intercept = 1,
  nsubsamples = 1000,
  nwarmup_samples = 1000,
  warm_start = 1
)

fqr(
  formula,
  data,
  tau = 0.5,
  se = T,
  smoothing_window = .Machine$double.eps,
  maxiter = 10000,
  beta_tol = 1e-16,
  check_tol = 1e-15 * nrow(data),
  nwarmup_samples = 1000,
  warm_start = 1,
  nsubsamples = 1000
)

Arguments

X
design matrix

y
outcome variable

tau
vector of target quantile(s)

se
whether to calculate standard errors

init_beta
initial betas for gradient descent (optional, default is random normal initial values)

smoothing_window
neighborhood around 0 to smooth with tilted least-squares loss function

maxiter
maximum number of allowed iterations for gradient descent

beta_tol
stopping criterion based on largest value of the gradient

check_tol
stopping criterion based on the change in the value of the check function between iterations

intercept
what column the intercept is, defaults to 1, 0 to indicate no intercept

nsubsamples
number of subsamples to use when calculating standard errors

nwarmup_samples
number of samples to use for warmup regression

warm_start
whether to run initial warmup regression or just default inits for full data gradient descent

formula
Regression formula

data
data to use when fitting regression

Details

This package performs quantile regression by approximating the check loss function with a least-squares loss function in a small neighbor around 0. Since the only point where the check function is not differentiable is at 0, this allows for first-order gradient descent methods to work.

This package uses "accelerated" gradient descent, which moves the future guess at the coefficients not only by a step size * the gradient, but also based on the prior momentum of the changes in the coefficient which leads to faster convergence. Gradient-based methods work at scale (both in terms of observations and dimension), and are much faster than the interior point algorithms in the quantreg package for large problems. Still, they are sometimes less exact for small problems.

The algorithm employs two early stopping rules: the `check_tol` argument stops based on the scaled change in the check function loss. The `beta_tol` argument stops based on the largest value of the gradient vector.

Before using the full dataset, the optimizer "warms up" on a random subset of the dataset. nwarmup_samples controls the size of that. `warm_start` is an integer which controls whether that happens at all (it is _strongly_ recommended).

Examples

fit <- fqr(area ~ peri, data = rock, tau = c(0.25, 0.5, 0.75))

# print coefficients & SEs
print(fit)

#> Tau:  0.25 
#>              Coefficient           SE
#> (Intercept) 5.213727e+03 499.87076412
#> peri        4.156585e-02   0.01308446
#> Tau:  0.5 
#>              Coefficient           SE
#> (Intercept) 7.349530e+03 4.085500e+02
#> peri        4.909627e-02 6.188583e-03
#> Tau:  0.75 
#>              Coefficient           SE
#> (Intercept) 8.738315e+03 3.911618e+02
#> peri        4.080978e-02 5.178668e-03

# grab coefficient vector
coef(fit)

#>              [,1]         [,2]         [,3]
#> [1,] 5.213727e+03 7.349530e+03 8.738315e+03
#> [2,] 4.156585e-02 4.909627e-02 4.080978e-02

# predict values
predict(fit)

#>        [,1]     [,2]     [,3]
#> 1  5329.775 7486.601 8852.252
#> 2  5375.526 7540.642 8897.171
#> 3  5377.108 7542.510 8898.724
#> 4  5374.559 7539.499 8896.221
#> 5  5377.852 7543.388 8899.454
#> 6  5380.413 7546.413 8901.968
#> 7  5394.362 7562.890 8915.664
#> 8  5394.320 7562.841 8915.623
#> 9  5366.774 7530.304 8888.578
#> 10 5342.525 7501.662 8864.770
#> 11 5399.944 7569.483 8921.145
#> 12 5379.419 7545.239 8900.992
#> 13 5381.509 7547.709 8903.045
#> 14 5359.958 7522.252 8881.885
#> 15 5379.964 7545.884 8901.528
#> 16 5364.573 7527.703 8886.416
#> 17 5405.290 7575.798 8926.393
#> 18 5412.729 7584.584 8933.697
#> 19 5415.913 7588.345 8936.823
#> 20 5399.918 7569.452 8921.119
#> 21 5356.246 7517.868 8878.241
#> 22 5394.669 7563.253 8915.966
#> 23 5408.989 7580.167 8930.025
#> 24 5359.974 7522.272 8881.902
#> 25 5295.919 7446.612 8819.012
#> 26 5271.061 7417.251 8794.606
#> 27 5293.377 7443.610 8816.516
#> 28 5279.627 7427.368 8803.015
#> 29 5290.674 7440.417 8813.862
#> 30 5265.255 7410.392 8788.905
#> 31 5285.559 7434.375 8808.840
#> 32 5274.457 7421.262 8797.940
#> 33 5273.032 7419.578 8796.541
#> 34 5254.894 7398.154 8778.732
#> 35 5269.873 7415.847 8793.439
#> 36 5274.457 7421.262 8797.940
#> 37 5270.951 7417.120 8794.498
#> 38 5233.526 7372.915 8757.753
#> 39 5263.168 7407.928 8786.856
#> 40 5282.101 7430.291 8805.445
#> 41 5252.863 7395.756 8776.739
#> 42 5226.556 7364.683 8750.910
#> 43 5261.349 7405.779 8785.070
#> 44 5308.518 7461.493 8831.381
#> 45 5262.530 7407.174 8786.230
#> 46 5238.576 7378.880 8762.711
#> 47 5274.242 7421.008 8797.729
#> 48 5275.477 7422.466 8798.941

# predict values with new data
predict(fit, newdata = head(rock))

#>       [,1]     [,2]     [,3]
#> 1 5329.775 7486.601 8852.252
#> 2 5375.526 7540.642 8897.171
#> 3 5377.108 7542.510 8898.724
#> 4 5374.559 7539.499 8896.221
#> 5 5377.852 7543.388 8899.454
#> 6 5380.413 7546.413 8901.968