Fast Quantile Regression

Fast Quantile Regression

fit_fqr(
  X,
  y,
  tau,
  se = T,
  init_beta = stats::rnorm(ncol(X)),
  smoothing_window = .Machine$double.eps,
  maxiter = 10000,
  beta_tol = 1e-16,
  check_tol = 1e-15 * nrow(X),
  intercept = 1,
  nsubsamples = 1000,
  nwarmup_samples = 1000,
  warm_start = 1
)

fqr(
  formula,
  data,
  tau = 0.5,
  se = T,
  smoothing_window = .Machine$double.eps,
  maxiter = 10000,
  beta_tol = 1e-16,
  check_tol = 1e-15 * nrow(data),
  nwarmup_samples = 1000,
  warm_start = 1,
  nsubsamples = 1000
)

Arguments

X
design matrix

y
outcome variable

tau
vector of target quantile(s)

se
whether to calculate standard errors

init_beta
initial betas for gradient descent (optional, default is random normal initial values)

smoothing_window
neighborhood around 0 to smooth with tilted least-squares loss function

maxiter
maximum number of allowed iterations for gradient descent

beta_tol
stopping criterion based on largest value of the gradient

check_tol
stopping criterion based on the change in the value of the check function between iterations

intercept
what column the intercept is, defaults to 1, 0 to indicate no intercept

nsubsamples
number of subsamples to use when calculating standard errors

nwarmup_samples
number of samples to use for warmup regression

warm_start
whether to run initial warmup regression or just default inits for full data gradient descent

formula
Regression formula

data
data to use when fitting regression

Details

This package performs quantile regression by approximating the check loss function with a least-squares loss function in a small neighbor around 0. Since the only point where the check function is not differentiable is at 0, this allows for first-order gradient descent methods to work.

This package uses "accelerated" gradient descent, which moves the future guess at the coefficients not only by a step size * the gradient, but also based on the prior momentum of the changes in the coefficient which leads to faster convergence. Gradient-based methods work at scale (both in terms of observations and dimension), and are much faster than the interior point algorithms in the quantreg package for large problems. Still, they are sometimes less exact for small problems.

The algorithm employs two early stopping rules: the `check_tol` argument stops based on the scaled change in the check function loss. The `beta_tol` argument stops based on the largest value of the gradient vector.

Before using the full dataset, the optimizer "warms up" on a random subset of the dataset. nwarmup_samples controls the size of that. `warm_start` is an integer which controls whether that happens at all (it is _strongly_ recommended).

Examples

fit <- fqr(area ~ peri, data = rock, tau = c(0.25, 0.5, 0.75))

# print coefficients & SEs
print(fit)

#> Tau:  0.25 
#>              Coefficient           SE
#> (Intercept) 5.213727e+03 498.40343227
#> peri        4.156585e-02   0.01486405
#> Tau:  0.5 
#>              Coefficient           SE
#> (Intercept) 7.420291e+03 4.085429e+02
#> peri        2.489165e-02 6.189918e-03
#> Tau:  0.75 
#>              Coefficient           SE
#> (Intercept) 8.824772e+03 3.911653e+02
#> peri        1.668214e-02 5.178325e-03

# grab coefficient vector
coef(fit)

#>              [,1]         [,2]         [,3]
#> [1,] 5.213727e+03 7.420291e+03 8.824772e+03
#> [2,] 4.156585e-02 2.489165e-02 1.668214e-02

# predict values
predict(fit)

#>        [,1]     [,2]     [,3]
#> 1  5329.775 7489.786 8871.346
#> 2  5375.526 7517.184 8889.708
#> 3  5377.108 7518.131 8890.343
#> 4  5374.559 7516.604 8889.320
#> 5  5377.852 7518.576 8890.642
#> 6  5380.413 7520.110 8891.669
#> 7  5394.362 7528.463 8897.268
#> 8  5394.320 7528.439 8897.251
#> 9  5366.774 7511.943 8886.196
#> 10 5342.525 7497.421 8876.464
#> 11 5399.944 7531.806 8899.508
#> 12 5379.419 7519.515 8891.271
#> 13 5381.509 7520.767 8892.110
#> 14 5359.958 7507.860 8883.460
#> 15 5379.964 7519.841 8891.490
#> 16 5364.573 7510.624 8885.312
#> 17 5405.290 7535.008 8901.654
#> 18 5412.729 7539.462 8904.639
#> 19 5415.913 7541.369 8905.917
#> 20 5399.918 7531.790 8899.498
#> 21 5356.246 7505.638 8881.970
#> 22 5394.669 7528.647 8897.391
#> 23 5408.989 7537.223 8903.138
#> 24 5359.974 7507.870 8883.467
#> 25 5295.919 7469.511 8857.759
#> 26 5271.061 7454.625 8847.782
#> 27 5293.377 7467.989 8856.739
#> 28 5279.627 7459.754 8851.220
#> 29 5290.674 7466.370 8855.654
#> 30 5265.255 7451.148 8845.452
#> 31 5285.559 7463.307 8853.601
#> 32 5274.457 7456.659 8849.145
#> 33 5273.032 7455.805 8848.573
#> 34 5254.894 7444.943 8841.293
#> 35 5269.873 7453.913 8847.305
#> 36 5274.457 7456.659 8849.145
#> 37 5270.951 7454.559 8847.738
#> 38 5233.526 7432.147 8832.718
#> 39 5263.168 7449.898 8844.614
#> 40 5282.101 7461.236 8852.213
#> 41 5252.863 7443.727 8840.479
#> 42 5226.556 7427.973 8829.920
#> 43 5261.349 7448.809 8843.884
#> 44 5308.518 7477.056 8862.815
#> 45 5262.530 7449.516 8844.358
#> 46 5238.576 7435.171 8834.744
#> 47 5274.242 7456.530 8849.059
#> 48 5275.477 7457.269 8849.554

# predict values with new data
predict(fit, newdata = head(rock))

#>       [,1]     [,2]     [,3]
#> 1 5329.775 7489.786 8871.346
#> 2 5375.526 7517.184 8889.708
#> 3 5377.108 7518.131 8890.343
#> 4 5374.559 7516.604 8889.320
#> 5 5377.852 7518.576 8890.642
#> 6 5380.413 7520.110 8891.669